Re-factoring JS into modular
